Update spectroscopy code

+++ C_Code/Webcam/Display/calibrations.cpp
@@ -22,11 +22,11 @@ void save_calibration(const std::vector<int>& data, const std::string& filename)
 }
 
 // Compute column averages
-std::vector<int> compute_column_average(const cv::Mat& frame) {
+std::vector<float> compute_column_average(const cv::Mat& frame) {
     int width = frame.cols;
     int height = frame.rows;
 
-    std::vector<int> column_data(width, 0);
+    std::vector<float> column_data(width, 0);
 
     for (int y = 0; y < height; y++) {
         const uint8_t* row = frame.ptr<uint8_t>(y);
@@ -43,9 +43,9 @@ std::vector<int> compute_column_average(const cv::Mat& frame) {
 }
 
 // Capture dark calibration
-std::vector<int> calibrate(cv::VideoCapture& cap, int frames) {
+std::vector<float> calibrate(cv::VideoCapture& cap, int frames) {
     cv::Mat frame;
-    std::vector<int> avg;
+    std::vector<float> avg;
 
     for (int i = 0; i < frames; i++) {
         cap >> frame;
@@ -53,7 +53,7 @@ std::vector<int> calibrate(cv::VideoCapture& cap, int frames) {
 
         cv::cvtColor(frame, frame, cv::COLOR_YUV2GRAY_YUYV);
 
-        std::vector<int> curr = compute_column_average(frame);
+        std::vector<float> curr = compute_column_average(frame);
 
         if (avg.empty()) {
             avg = curr;
